Commit
The changes I made in Github were creating my own GitHub account and adding Website files to a local repository folder. GitHub is a central place where all project files, code, and documentation are stored, allowing anyone to view the code, or private. The other change I made was that I had to download a copy of the repository so It could be downloaded.